devel/py-gitpython: use the right gitdb dependency
gitdb2 is the "old" gitdb; it is still at version 2.0.
On PyPy it is listed as a mirror of gitdb and has version 4.0,
but gitpython wants gitdb 4.0.7 or later (as stated in its
`requires.txt`). While gitpython **builds** against the
older version, at runtime anything using setuptools to
require gitpython will fall over (because the installed
gitdb2 is way too old).

Switch to the right, modern dependency.

PORTREVISION bumped because of this make-it-actually-work
dependency change.