s3:modules:crossrename -validate sizelimit
The `crossrename:sizelimit` parameter is used to determine
the value of module_sizelimit via lp_int(), which can
return a zero in case of parsing error.

Add validation that parameter contains only numeric
digits. In case of invalid parameter, then deny access to
share.

Zero value for this parameter makes no sense (basically turns
module into a no-op). Additionally, it is useful to provide
users with a configuration option to remove the sizelimit
entirely. For this reason, introduce new behavior for case
where module_sizelimit is zero, which skips the sizelimit
check.