NAS-113336 / 12.0 / s3:modules:full_audit - back out sanity checks (#1102)
The upgrade from 4.12 to 4.13 caused an internal check
in vfs_full_audit to fail. This PR backs out a patch
we made to protect users from logging misconfiguration
on Samba version upgrades. The practical impact of
removing the patch will be that some users may have
samba fall back to logging every VFS operation in
upgrading from 12.0-U6.1 to 12.0-U7, but this is preferable
to having a broken full_audit module.