devel/intel-pcm: Update to 202112
There's a new diagnostic in the build:

"The compiler is too old (g++ 6 or below). Parsing events from
01.org/perfmon won't be supported."

despite the fact we're using the latest clang. I'll investigate and see
what other dependencies are needed (since it's testing for the presence
of other libraries). I'll update with the improved functionality when
that's been sorted.

Sponsored by:		Netflix