Fix validation for platform-specific field lengths
Netatalk writes sizeof(dev_t) and sizeof(ino_t) into its
metadata xattr and then uses this in validation checks whether
this value is 8. Data written in FreeBSD prior to 12 (or
other platforms with 32-bit inode or dev_t will fail this check).

Add additional bool `platform_specific` to indicate a value that
can be a fixed length of either 8 or 4.