net/samba - fix winmsa creator handling
In presence of creator-owner or creator-group SID in the DACL
from parent directory, we need to retrieve the owner SID and
group SID of the target file before the permissions change.
We do this by converting uid and gid from stat into SIDs and
then passing to the function to generate a child security
descriptor. This child security descriptor is then used to
set an ACL on the file.

https://github.com/truenas/samba/pull/217