modules:ixnas - normalize DOS mode config
Some of the kernel DOS mode behavior isn't standardized
between Linux and FreeBSD and we have a variety of users
who are migrating data between CORE and SCALE freely.

This switches to using the non-optimized path for DOS
modes. The performance impact of DOSATTR lookups should
be minimal these days now that we've switched to using
SA-BASED xattrs.

This PR also brings in a fix for an edge case with
MacOS SMB clients where attempts to remove the DOS
readonly flag could be ignored.