net/samba - Add VFS endpoint for optmized lookup (#1272)
Provide non-default fast-path for handle lookups within Samba's vfs.
The fast-path in this case is an RB-tree keyed on device-inode pairs
containing fhandle_t handle that can be used to open a file in a race-
free manner without having to lookup vnode. Cache is populated during
directory listing. When we are able to successfully open from the
fhandle cache, we can skip symlink safety checks which results in
significant performance improvement for workloads involving large
numbers of files.