net/samba - fix ACL inheritance in vfs_winmsa
This commit brings in changes to libzfsacl that provide API for
generating inherited ACL from arbitrary ACL, and then refactors
vfs_winmsa to use the new API in lieu of passing through Samba's
VFS. This is significantly more performant and stable, but comes
at the expense of being suitable only for ZFS with acltype set
to NFSv4 and will not stack properly with modules that write
metadata into separate files (e.g. vfs_fruit). This is an
acceptable compromise though because the module is not available
via the webui or middleware for typical end-users.